01-Ai_agents_first/04_streaming/streamed-chatbot/pyproject.toml
[project]
name = "streamed-chatbot"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "openai-agents>=0.0.16",
]
